+++ src/bayestar_moc.h
@@ -40,7 +40,7 @@ int8_t uniq2nest64(uint64_t uniq, uint64_t *nest);
 
 void uniq2ang64(uint64_t uniq, double *theta, double *phi);
 
-void *moc_rasterize64(const void *pixels, size_t offset, size_t itemsize, size_t len, size_t *npix);
+void *moc_rasterize64(const void *pixels, size_t offset, size_t itemsize, size_t len, size_t *npix, int8_t order);
 
 #endif /* __cplusplus */